center_sample_radius=1.0

+++ models/detector/fcos/matcher.py
@@ -214,9 +214,9 @@ class Matcher(object):
                 gt_anchors_reg_deltas_i = torch.zeros([anchors_over_all_feature_maps.shape[0], 4], device=device)
                 gt_centerness_i = torch.zeros(anchors_over_all_feature_maps.shape[0], device=device)
 
-                gt_classes.append(tgt_cls_i)
-                gt_anchors_deltas.append(gt_anchors_reg_deltas_i)
-                gt_centerness.append(gt_centerness_i)
+                gt_classes.append(tgt_cls_i.long())
+                gt_anchors_deltas.append(gt_anchors_reg_deltas_i.float())
+                gt_centerness.append(gt_centerness_i.float())
 
         # [B, M], [B, M, 4], [B, M]
         return torch.stack(gt_classes), torch.stack(gt_anchors_deltas), torch.stack(gt_centerness)
